+++ lib/load.js
@@ -2,9 +2,11 @@ var path        = require('path')
   , fs          = require('fs')
   , merge       = require('deeply')
   , typeOf      = require('precise-typeof')
-  , stripBOM    = require('strip-bom')
+  , stripBOM    = require('stripbom')
+  , applyHooks  = require('./apply_hooks.js')
   , getCacheKey = require('./get_cache_key.js')
   , getFiles    = require('./get_files.js')
+  , mergeLayers = require('./merge_layers.js')
   , resolveExts = require('./resolve_exts.js')
   ;
 
@@ -116,7 +118,7 @@ function loadFiles(dirs, files)
 }
 
 /**
- * Loads and parses provided file (synchronious).
+ * Loads and parses provided file (synchronous).
  *
  * @param   {string} file - absolute path to the file
  * @param   {function} parser - function to parse provided content and return config object
@@ -129,7 +131,7 @@ function loadContent(file, parser)
   try
   {
     content = stripBOM(fs.readFileSync(file, {encoding: 'utf8'}));
-    // provide filepath as the second argument for complex parsing,
+    // provide file path as the second argument for complex parsing,
     // also it matches `module._compile` nodejs API.
     // Note: JSON.parse accepts two arguments, but ignores anything
     // but function on the second place, so it's safe to pass a filename
@@ -142,57 +144,4 @@ function loadContent(file, parser)
   }
 
   return config;
-}
-
-/**
- * Applies matched hooks
- *
- * @param   {object} config - config object to apply hooks to
- * @param   {string} filename - base filename to match hooks against
- * @returns {object} - modified config object
- */
-function applyHooks(config, filename)
-{
-  Object.keys(this.hooks).forEach(function(hook)
-  {
-    // in order to match hook should either the same length
-    // as the filename or smaller
-    if (filename.substr(0, hook.length) === hook)
-    {
-      config = this.hooks[hook](config);
-    }
-  }.bind(this));
-
-  return config;
-}
-
-/**
- * Merges provided layers into a single config object,
- * respecting order of the layers
- *
- * @param   {array} layers - list of config objects
- * @returns {object} - single config object
- */
-function mergeLayers(layers)
-{
-  var _instance = this
-    , result    = {}
-    ;
-
-  layers.forEach(function(layer)
-  {
-    layer.exts.forEach(function(ext)
-    {
-      ext.dirs.forEach(function(cfg)
-      {
-        // have customizable's array merge function
-        result = merge.call({
-          useCustomAdapters: merge.behaviors.useCustomAdapters,
-          'array': _instance.arrayMerge
-        }, result, cfg.config);
-      });
-    });
-  });
-
-  return result;
 }
