+++ package.json
@@ -41,7 +41,10 @@
     "json5",
     "properties",
     "toml",
-    "yaml"
+    "yaml",
+    "sdk",
+    "toolkit",
+    "diy"
   ],
   "author": "Alex Indigo <iam@alexindigo.com>",
   "license": "MIT",
@@ -52,23 +55,23 @@
     "compare-property": "^2.0.0",
     "deeply": "^2.0.1",
     "precise-typeof": "^1.0.2",
-    "strip-bom": "^2.0.0"
+    "stripbom": "^3.0.0"
   },
   "devDependencies": {
     "coffee-script": "^1.10.0",
-    "coveralls": "^2.11.9",
+    "coveralls": "^2.11.12",
     "cson": "^3.0.2",
-    "eslint": "^2.9.0",
+    "eslint": "^2.13.1",
     "hjson": "^1.8.4",
     "ini": "^1.3.4",
-    "js-yaml": "^3.6.0",
+    "js-yaml": "^3.6.1",
     "json5": "^0.5.0",
-    "nyc": "^6.4.2",
-    "pre-commit": "^1.1.2",
+    "nyc": "^7.1.0",
+    "pre-commit": "^1.1.3",
     "properties": "^1.2.1",
-    "sinon": "^1.17.4",
+    "sinon": "^1.17.5",
     "tap-spec": "^4.1.1",
-    "tape": "^4.5.1",
+    "tape": "^4.6.0",
     "toml": "^2.3.0"
   }
 }
